+ Created pop up menu and charater ınformation panel

+++ Assets/Scripts/Managers/UIManager.cs
@@ -16,7 +16,7 @@ namespace DarkJimmy
         private Stack<Menu> _stack = new Stack<Menu>();
         private readonly Vector2 canvasResolition =  new Vector2(2960,1440);
 
-        public delegate void UpdateState(Stats state,int amount);  
+        public delegate void UpdateState(Stats state,float amount);  
         public UpdateState updateState;
         public UpdateState addCollectable;
         public override void Awake()
